# [HWASAN] Use sign extension in memToShadow() and untagPointer()