This allows the shift in untagPointer() to be folded with the shift in memToShadow(), and works for both user and kernel HWASAN. On AArch64, the sequence collapses to a single sbfx instruction, and on RISC-V it avoids consuming a register to hold the shifted mask byte.

This is a backward-compatible change because (for a constant ShadowBase) it only affects the top bits of the pointer returned from memToShadow(), which are ignored by hardware anyway. Note that the AArch64 outline assembly already uses a sbfx instruction for these combined operations.

Additionally use the original tagged pointer to compute the inline tag address, to avoid needing to keep the untagged pointer live.