@fmayer 2024-08-14
Tests failed in presubmit:

```
Failed Tests (1):
  LLVM :: Instrumentation/HWAddressSanitizer/globals-access.ll
```